Rework run_kontact.sh to prevent desktop and dbus bypasses
If a fresh install was first executed via CLI and later with the desktop
file it would seem like all data is gone because a different Akonadi
instance would be used due to missing AKONADI_INSTANCE=flatpak export.

Revamp and improve the run_kontact.sh script to be a generic wrapper for
anything launched is the flatpak. This ensures variables are set
properly. Akonadi is also no longer started and shut down excessively.